Moved foreach
